pyproject.toml
[project]
name = "hifisuperstar"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py==2.7.1",
    "duckduckgo-search==8.1.1",
    "Flask==3.1.3",
    "requests==2.33.1",
    "spotipy==2.26.0",
    "validators==0.35.0",
    "yt-dlp==2026.3.17",
    "GitPython==3.1.46",
    "numpy==2.4.4",
    "pynacl>=1.6.2",
    "davey>=0.1.5",
]
